Finish and test function of server and client

+++ src/sysu/java/server/host/actions/RemoveFileAction.java
@@ -1,5 +1,6 @@
 package sysu.java.server.host.actions;
 
+import sysu.java.Utils;
 import sysu.java.server.host.ServerCommands;
 
 import java.io.IOException;
@@ -11,17 +12,25 @@ import java.nio.file.Paths;
 public class RemoveFileAction extends Action {
 
     @Override
-    public void execute(int length, InputStream input) {
+    public boolean execute(int length, InputStream input) {
         try {
-            var path = Paths.get(host.getBasePath(), new String(input.readNBytes(length), StandardCharsets.UTF_8)).toAbsolutePath();
+            var pathStr = new String(input.readNBytes(length), StandardCharsets.UTF_8);
+            if (!Utils.checkPath(pathStr) || pathStr.isBlank()) {
+                writeFailure("Illegal path");
+                return false;
+            }
+            var path = Paths.get(host.getBasePath(), pathStr).toAbsolutePath();
             var file = path.toFile();
             if (file.isFile() && file.exists() && file.delete()) {
                 writeSuccess("File removed successfully");
+                return true;
             } else {
                 writeSuccess("File removed failed");
+                return false;
             }
         } catch (IOException e) {
             e.printStackTrace();
+            return false;
         }
     }
 }
