+++ src/sysu/java/server/host/actions/RemoveFolderAction.java
@@ -11,18 +11,26 @@ import java.nio.file.Paths;
 @RegisterAction(command = ServerCommands.RemoveFolder)
 public class RemoveFolderAction extends Action {
     @Override
-    public void execute(int length, InputStream input) {
+    public boolean execute(int length, InputStream input) {
         try {
-            var path = Paths.get(host.getBasePath(), new String(input.readNBytes(length), StandardCharsets.UTF_8)).toAbsolutePath();
+            var pathStr = new String(input.readNBytes(length), StandardCharsets.UTF_8);
+            if (!Utils.checkPath(pathStr) || pathStr.isBlank()) {
+                writeFailure("Illegal path");
+                return false;
+            }
+            var path = Paths.get(host.getBasePath(), pathStr).toAbsolutePath();
             var file = path.toFile();
             if (file.isDirectory() && file.exists()) {
                 Utils.deleteDirectory(file);
                 writeSuccess("Folder removed successfully");
+                return true;
             } else {
-                writeSuccess("Folder removed failed");
+                writeFailure("Folder removed failed");
+                return false;
             }
         } catch (IOException e) {
             e.printStackTrace();
+            return false;
         }
     }
 }
